requirements.txt
apipkg==1.5
atomicwrites==1.3.0
attrs==19.3.0
click==7.1.1
colorama==0.4.3
coverage==5.0.4
eventlet~=0.25.2
execnet==1.7.1
Flask==1.1.2
Flask-SocketIO==4.2.1
flask-sqlalchemy==2.4.1
humanfriendly==8.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
package-name==0.1
packaging==20.3
passlib==1.7.2
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pyreadline==2.1
pytest==5.4.1
pytest-cov==2.8.1
pytest-dependency==0.5.1
pytest-forked==1.1.3
pytest-xdist==1.31.0
python-engineio==3.12.1
python-socketio==4.5.1
pyyaml==5.3.1
six==1.14.0
sqlalchemy==1.3.15
wcwidth==0.1.9
Werkzeug==1.0.0
zipp==3.1.0
colorlog==4.1.0
python-memcached==1.59
flask_sessionstore
flask_session